+++ vembrane/ann_types.py
@@ -23,6 +23,8 @@ class NoValue:
     def __bool__(self):
         return False
 
+    def __repr__(self):
+        return "."
 
 NA = NoValue()
