+++ vembrane/common.py
@@ -1,6 +1,11 @@
 import ast
+from typing import Dict, List, Tuple
+
 from vembrane.errors import InvalidExpression
 
+from pysam.libcbcf import (
+    VariantHeader,
+)
 
 def check_filter_expression(expression: str,) -> str:
     if ".__" in expression:
@@ -16,3 +21,16 @@ def check_filter_expression(expression: str,) -> str:
         raise InvalidExpression(
             expression, "The expression has to be syntactically correct."
         )
+
+
+def get_annotation_keys(header: VariantHeader, ann_key: str) -> List[str]:
+    separator = "'"
+    for rec in header.records:
+        if rec.key == "VEP":
+            separator = ":"
+            continue
+        if rec.get("ID") == ann_key:
+            return list(
+                map(str.strip, rec.get("Description").split(separator)[1].split("|"))
+            )
+    return []
